:recycle: CalendarBoardのリファクタ / ロジックをutilsとして分割
