+++ front/src/components/CalendarBoard/styles.ts
@@ -8,9 +8,3 @@ export const gridStyle = {
   borderLeft: '1px solid #ccc',
   borderTop: '1px solid #ccc',
 };
-
-export const elementStyle = {
-  borderRight: '1px solid #ccc',
-  borderBottom: '1px solid #ccc',
-  height: '18vh',
-};
